Removed the need for the user to serialize the type for polymorphic serialization, serialized types are now handled
Updated the comments for the Serializer and the SerializeConstruct for what each function is required required to do.
Removed the requirement for serializing arrays and strings from the serializer until a common interface has been solved.
Removed the requirement that a serializer must open / close a file and functions related to so, and instead opted for a dump function